mi_app/serializers/categoria_serializer.py
from rest_framework import serializers
from mi_app.models import Categoria


class CategoriaSerializer(serializers.ModelSerializer):
    class Meta:
        model = Categoria
        fields = '__all__'