requirements.txt
Django==5.2
djangorestframework==3.15.0
django-filter==23.2
psycopg2-binary==2.9.9
